+++ src/server/BPMNServer.ts
@@ -2,7 +2,7 @@
 import { Logger } from '../common/Logger';
 
 
-import { IConfiguration, ILogger, DataStore , IAppDelegate, IBPMNServer, IDataStore,ICacheManager,IModelsDatastore, IUserService} from '../';
+import { IConfiguration, ILogger, DataStore , IAppDelegate, IBPMNServer, IDataStore,ICacheManager,IModelsDatastore} from '../';
 import { Engine } from './Engine';
 import { CacheManager } from './CacheManager';
 import { Cron } from './Cron';
@@ -50,7 +50,6 @@ class BPMNServer implements IBPMNServer {
 	dataStore: IDataStore;
 	cache: ICacheManager;
 	cron: Cron;
-	userService: IUserService;
 	error;
 
 	private static instance: BPMNServer;
@@ -76,8 +75,7 @@ class BPMNServer implements IBPMNServer {
 		this.dataStore = configuration.dataStore(this);
 		this.definitions = configuration.definitions(this);
 		this.appDelegate = configuration.appDelegate(this);
-		this.userService=configuration.userService(this);
-
+	
 		//this.acl = new ACL(this);
 		//this.iam = new IAM(this);
 		console.log("bpmn-server version " + BPMNServer.getVersion());